Lint messages on save to fix HTML tags
Some messages would not be rendered as valid on the client, but when
added to chat history messages, invalid HTML or unclosed tags would end
up messing up the formatting for the rest of the replay message.

This uses tidylib in order to validate good HTML or escape bad HTML
text messages. Fixes #5.